move middleware, handle 401 error on app selection
